move story.py into views package
