+++ game/views/story.py
